feat(rhi): descriptor 버전을 중립 어휘로 올려 draw 신원 세 축을 캡처에 싣는다 — PBR-W0
W0 의 남은 항목은 "draw-level sampler identity·descriptor generation·resolved
PSO key 를 capture manifest 에" 였다. 재 보니 셋 중 둘은 이미 있었다 —
sealLedger.<pass>.bindings 가 sealHash 마다 pipelineId 와 samplerIdentity 를
적고 draws[].seal.hash 로 조인된다(실측 draw 10 · 조인 10/10 · 누락 0).
값을 draw 항목에 복사하지 않은 것은 설계다: "같은 밀봉을 공유하는 draw 는 같은
바인딩을 쓴다"가 W8 의 불변식이라, 복사하면 그 불변식이 표에서 사라진다.

descriptor generation 만 제품에 없었다. Binding::descriptorVersion 은 writer 가
0 인 죽은 칸이었다 — 선언만 있고 아무도 채우지 않았고 방출되지도 않았다. 값
자체(RHIDescriptorVersionHandle::ToToken)는 백엔드 중립인데 **꺼내는 어휘가
DX12 에만** 있었다. 어휘 구멍이 백엔드 비대칭을 만든 자리다.

- VulkanDescriptorPoolRecycler::GetCurrentVersionToken 으로 DX12 와 짝을 맞췄다.
- IRenderDeviceServices::GetDescriptorVersionToken 으로 중립 계층에 올렸다.
  구현 셋(DX12·Vulkan·자가 검사 대역) 중 대역은 0 을 돌려주고, 0 의 뜻은
  "기록 중인 버전 없음" 으로 인터페이스가 규정한다.
- 두 패스가 바인딩 기록에 적고 캡처가 방출한다.

★ 폭을 uint32 에서 uint64 로 넓힌 것이 결정적이었다. 토큰은 generation << 32 |
slot 이고 실측값이 17179869185(= 0x4_0000_0001) 이므로 옛 폭 그대로였다면 윗
32비트가 잘려 1 로 보였을 것이다 — 죽은 칸이라 아무도 몰랐다.

게이트(Assert-Capture)는 둘만 단정한다: 모든 draw 의 seal.hash 가 **자기 라우트**
장부에서 바인딩을 찾는 것, 그리고 세 축이 0 이 아닌 것. 변이 폭(distinct pso
1~2 · sampler 1 · descriptorVersion 1)은 매 실행 출력에 남기되 판정에서 뺐다 —
sampler 가 1 인 것은 지금 pass-global 이라 옳고 재질별로 가르는 것은 W7 이다.
같은 이유로 descriptorVersion 을 바인딩 신원(operator==)에 넣지 않았다.

변이 증명: DX12 토큰 접근자를 0 으로 바꾸니 예고한 대로 첫 캡처에서
"gbuffer binding has descriptorVersion=0 for seal 5568184615101585538" 로 붉었고
변이 폭을 적는 식별 줄은 찍히기 전에 멈췄다. 복원 뒤 세 캡처 전부 조인 성립.

남은 것: W0 의 fixture 축(여덟 중 셋만 서 있다)과, forward 라우트 draw 가 있는
fixture 가 없어 라우트별 조인이 아직 자극된 적 없다는 것.

+++ Engine/RenderEngine/Render/Graph/EnhancedDrawSealLedger.h
@@ -105,7 +105,22 @@ struct EnhancedDrawSealLedger
         std::uint64_t textureDigest{};
         std::uint64_t samplerIdentity{};
         std::uint32_t pipelineId{};
-        std::uint32_t descriptorVersion{};
+
+        /// W0 — 이 바인딩이 어느 transient descriptor 버전에서 잘려 나왔는가.
+        /// `RHIDescriptorVersionHandle::ToToken()` 값이고 0 은 "버전 없음"이다.
+        ///
+        /// ★ 이 칸은 오래 **writer 가 0 이었다** — 선언만 있고 아무도 채우지
+        ///   않았으며 방출되지도 않았다. 값을 꺼낼 어휘가 DX12 에만 있었기
+        ///   때문이다(`IRenderDeviceServices::GetDescriptorVersionToken` 이 그것을
+        ///   중립으로 올렸다). 폭도 uint32 였는데 토큰은 generation<<32|slot 이라
+        ///   윗 32비트가 통째로 잘려 나갔을 자리다.
+        ///
+        /// ★ **아직 신원에 넣지 않는다.** 아래 `operator==` 는 이 값을 보지
+        ///   않는다 — 한 프레임에 기록이 둘 이상이면 같은 밀봉이 서로 다른
+        ///   버전에서 잘릴 수 있고, 그때 신원으로 쓰면 거짓 충돌이 된다.
+        ///   먼저 캡처로 실제 분포를 재고, 프레임당 하나임이 확인되면 그때
+        ///   신원으로 올린다. 재기 전에 판정하지 않는다.
+        std::uint64_t descriptorVersion{};
 
         bool operator==(const Binding& other) const noexcept
         {
